requirements.txt
# Core pipeline
python-dotenv>=1.0
requests>=2.28
numpy>=1.23
sentence-transformers>=2.2

# LLM & Classification
# (Groq is accessed via REST API, no SDK needed)

# Scheduling
APScheduler>=3.10

# Digest & Email
markdown2>=2.4

# Dashboard
streamlit>=1.28

# Testing
pytest>=7.0
